Extracted common tests for bonding abstract contract
